divide events on newMessage and updatedMessage
