+++ app/Events/UpdatedMessageEvent.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace App\Events;
+
+class UpdatedMessageEvent extends MessageEvent
+{
+    public $socketEvent = 'updatedMessage';
+}
